Misc work to support nRF51/52 chips: o Move Kconfig to config/ o Splitting out chip selection into its own file. o Supporting using source directive in Kconfig. Pass the .config file as a environment variable and execute kconfig from config/ output_stream.h: Remove unused code.

+++ src/io/output_stream.cpp
@@ -30,7 +30,7 @@ size_t formatted_output_stream::println(const char *format, ...)
     va_end(args);
 
     auto written = write(reinterpret_cast<uint8_t *>(buffer), count);
-    if (written == count) {
+    if (written == static_cast<size_t>(count)) {
         uint8_t c = '\n';
         written += write(&c, 1);
     }
